requirements.txt
fastapi==0.109.0
uvicorn[standard]==0.27.0
jinja2==3.1.3
python-multipart==0.0.6
httpx==0.26.0
pydantic==2.5.3
openai==1.54.3
anthropic==0.39.0
python-dotenv==1.0.0

